+++ database/migrations/2018_06_09_145733_create_companies_table.php
@@ -14,7 +14,7 @@ class CreateCompaniesTable extends Migration
     public function up()
     {
         Schema::create('companies', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('business_name', 150);
             $table->string('tin', 16)->nullable();
             $table->string('email', 150)->nullable();
@@ -22,10 +22,10 @@ class CreateCompaniesTable extends Migration
             $table->string('phone', 12)->nullable();
             $table->string('mobile', 12)->nullable();
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')
                 ->on('users')->onDelete('cascade')->onUpdate('cascade');
-                
+
             $table->timestamps();
         });
     }
