+++ database/migrations/2018_06_09_145848_create_vehicles_table.php
@@ -14,16 +14,16 @@ class CreateVehiclesTable extends Migration
     public function up()
     {
         Schema::create('vehicles', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('registration', 7)->unique();
             $table->string('brand')->nullable();
             $table->string('color')->nullable();
-        
+
             $table->integer('vehicle_type_id')->unsigned();
             $table->foreign('vehicle_type_id')->references('id')
                 ->on('vehicle_types')->onDelete('cascade')->onUpdate('cascade');
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')
                 ->on('users')->onDelete('cascade')->onUpdate('cascade');
