+++ database/migrations/2018_06_09_150053_create_services_table.php
@@ -14,15 +14,15 @@ class CreateServicesTable extends Migration
     public function up()
     {
         Schema::create('services', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('description')->unique();
             $table->decimal('price', 10, 2);
             $table->boolean('status')->default(true);
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')
                 ->on('users')->onDelete('cascade')->onUpdate('cascade');
-            
+
             $table->timestamps();
         });
     }
