+++ database/migrations/2018_07_11_170449_create_guest_vehicle_table.php
@@ -15,11 +15,11 @@ class CreateGuestVehicleTable extends Migration
     {
         Schema::create('guest_vehicle', function (Blueprint $table) {
             $table->bigInteger('guest_id')->unsigned();
-            $table->integer('vehicle_id')->unsigned();
+            $table->bigInteger('vehicle_id')->unsigned();
             $table->datetime('created_at')->nullable();
-            $table->integer('invoice_id')->nullable();
+            $table->bigInteger('invoice_id')->nullable();
             $table->string('other')->nullable();
-            
+
             $table->foreign('guest_id')->references('id')->on('guests')
                 ->onUpdate('cascade')->onDelete('cascade');
             $table->foreign('vehicle_id')->references('id')->on('vehicles')
