+++ database/migrations/2018_07_21_085021_create_guest_room_table.php
@@ -15,9 +15,9 @@ class CreateGuestRoomTable extends Migration
     {
         Schema::create('guest_room', function (Blueprint $table) {
             $table->bigInteger('guest_id')->unsigned();
-            $table->integer('room_id')->unsigned();
+            $table->bigInteger('room_id')->unsigned();
             $table->bigInteger('invoice_id');
-            
+
             $table->foreign('guest_id')->references('id')->on('guests')
                 ->onUpdate('cascade')->onDelete('cascade');
             $table->foreign('room_id')->references('id')->on('rooms')
